# Faster palette