--- a/common/src/main/java/us/myles/ViaVersion/api/type/types/version/ChunkSectionType1_13.java
+++ b/common/src/main/java/us/myles/ViaVersion/api/type/types/version/ChunkSectionType1_13.java
@@ -1,6 +1,5 @@
 package us.myles.ViaVersion.api.type.types.version;
 
-import com.google.common.collect.BiMap;
 import io.netty.buffer.ByteBuf;
 import us.myles.ViaVersion.api.minecraft.chunks.ChunkSection;
 import us.myles.ViaVersion.api.type.Type;
@@ -14,8 +13,6 @@ public class ChunkSectionType1_13 extends Type<ChunkSection> {
     @Override
     public ChunkSection read(ByteBuf buffer) throws Exception {
         ChunkSection chunkSection = new ChunkSection();
-        BiMap<Integer, Integer> palette = chunkSection.getPalette();
-        palette.clear();
 
         // Reaad bits per block
         int bitsPerBlock = buffer.readUnsignedByte();
@@ -32,8 +29,9 @@ public class ChunkSectionType1_13 extends Type<ChunkSection> {
         }
         int paletteLength = bitsPerBlock == 14 ? 0 : Type.VAR_INT.read(buffer);
         // Read palette
+        chunkSection.clearPalette();
         for (int i = 0; i < paletteLength; i++) {
-            palette.put(Type.VAR_INT.read(buffer), palette.size());
+            chunkSection.addPaletteEntry(Type.VAR_INT.read(buffer));
         }
 
         // Read blocks
@@ -68,10 +66,8 @@ public class ChunkSectionType1_13 extends Type<ChunkSection> {
 
     @Override
     public void write(ByteBuf buffer, ChunkSection chunkSection) throws Exception {
-        BiMap<Integer, Integer> palette = chunkSection.getPalette();
-
         int bitsPerBlock = 4;
-        while (palette.size() > 1 << bitsPerBlock) {
+        while (chunkSection.getPaletteSize() > 1 << bitsPerBlock) {
             bitsPerBlock += 1;
         }
 
@@ -85,9 +81,9 @@ public class ChunkSectionType1_13 extends Type<ChunkSection> {
 
         // Write pallet (or not)
         if (bitsPerBlock != 14) {
-            Type.VAR_INT.write(buffer, palette.size());
-            for (int i = 0; i < palette.size(); i++) {
-                Type.VAR_INT.write(buffer, palette.inverse().get(i));
+            Type.VAR_INT.write(buffer, chunkSection.getPaletteSize());
+            for (int i = 0; i < chunkSection.getPaletteSize(); i++) {
+                Type.VAR_INT.write(buffer, chunkSection.getPaletteEntry(i));
             }
         }