--- a/common/src/main/java/us/myles/ViaVersion/protocols/protocol1_13_1to1_13/packets/WorldPackets.java
+++ b/common/src/main/java/us/myles/ViaVersion/protocols/protocol1_13_1to1_13/packets/WorldPackets.java
@@ -1,6 +1,5 @@
 package us.myles.ViaVersion.protocols.protocol1_13_1to1_13.packets;
 
-import com.google.common.collect.BiMap;
 import us.myles.ViaVersion.api.PacketWrapper;
 import us.myles.ViaVersion.api.minecraft.BlockChangeRecord;
 import us.myles.ViaVersion.api.minecraft.chunks.Chunk;
@@ -29,9 +28,8 @@ public class WorldPackets {
 
                         for (ChunkSection section : chunk.getSections()) {
                             if (section == null) continue;
-                            BiMap<Integer, Integer> inverse = section.getPalette().inverse();
-                            for (int i = 0; i < inverse.size(); i++) {
-                                inverse.put(i, Protocol1_13_1To1_13.getNewBlockStateId(inverse.get(i)));
+                            for (int i = 0; i < section.getPaletteSize(); i++) {
+                                section.setPaletteEntry(i, Protocol1_13_1To1_13.getNewBlockStateId(section.getPaletteEntry(i)));
                             }
                         }
                     }