--- a/common/src/main/java/us/myles/ViaVersion/protocols/protocol1_9to1_8/types/Chunk1_9to1_8Type.java
+++ b/common/src/main/java/us/myles/ViaVersion/protocols/protocol1_9to1_8/types/Chunk1_9to1_8Type.java
@@ -89,8 +89,8 @@ public class Chunk1_9to1_8Type extends PartialType<Chunk, ClientChunks> {
             ChunkSection section = Types1_8.CHUNK_SECTION.read(input);
             sections[i] = section;
 
-            if (replacePistons && section.getPalette().contains(36)) {
-                section.getPalette().set(section.getPalette().indexOf(36), replacementId);
+            if (replacePistons && section.getPalette().containsKey(36)) {
+                section.getPalette().put(replacementId, section.getPalette().remove(36));
             }
         }